Extract sync apply helpers
