+++ tests/test-cashu-wallet.js
@@ -55,7 +55,7 @@ const walletSrc = await fetchWithRetry('js/cashu-wallet.js');
 const discoverySrc = await fetchWithRetry('js/nostr-discovery.js');
 const apiSrc = await fetchWithRetry('js/api.js');
 const ppSrc = await fetchWithRetry('js/provider-panels.js');
-const syncSrc = await fetchWithRetry('js/sync.js');
+const syncApplySrc = await fetchWithRetry('js/sync-apply.js');
 const syncPayloadSrc = await fetchWithRetry('js/sync-payload.js');
 const cryptoSrc = await fetchWithRetry('js/crypto.js');
 const backupSrc = await fetchWithRetry('js/backup.js');
@@ -215,7 +215,7 @@ console.log('11. Sync Integration');
 assert('Wallet mnemonic in AI_SETTINGS_KEYS', syncPayloadSrc.includes("'labcharts-cashu-wallet-mnemonic'"));
 assert('Wallet mint in AI_SETTINGS_KEYS', syncPayloadSrc.includes("'labcharts-cashu-wallet-mint'"));
 assert('Node URL in AI_SETTINGS_KEYS', syncPayloadSrc.includes("'labcharts-routstr-node'"));
-assert('Mnemonic in ENCRYPTED_AI_KEYS', syncSrc.includes("'labcharts-cashu-wallet-mnemonic'"));
+assert('Mnemonic in ENCRYPTED_AI_KEYS', syncApplySrc.includes("'labcharts-cashu-wallet-mnemonic'"));
 assert('Wallet keys in GLOBAL_SETTINGS_KEYS', backupSrc.includes("'labcharts-cashu-wallet-mint'") && backupSrc.includes("'labcharts-routstr-node'"));
 
 // ═══════════════════════════════════════
