+++ tests/test-export-import.js
@@ -417,11 +417,11 @@ return (async function() {
   else localStorage.removeItem('labcharts-custom-model');
 
   // Sync also picks them up (cross-device parity)
-  const syncSrc = await fetch('/js/sync.js').then(r => r.text());
   const syncPayloadSrc = await fetch('/js/sync-payload.js').then(r => r.text());
+  const syncApplySrc = await fetch('/js/sync-apply.js').then(r => r.text());
   assert('AI_SETTINGS_KEYS includes labcharts-custom-key', /AI_SETTINGS_KEYS[\s\S]{0,800}labcharts-custom-key/.test(syncPayloadSrc));
   assert('AI_SETTINGS_KEYS includes labcharts-custom-url', /AI_SETTINGS_KEYS[\s\S]{0,800}labcharts-custom-url/.test(syncPayloadSrc));
-  assert('ENCRYPTED_AI_KEYS includes labcharts-custom-key', /ENCRYPTED_AI_KEYS[\s\S]{0,400}labcharts-custom-key/.test(syncSrc));
+  assert('ENCRYPTED_AI_KEYS includes labcharts-custom-key', /ENCRYPTED_AI_KEYS[\s\S]{0,400}labcharts-custom-key/.test(syncApplySrc));
 
   // ═══════════════════════════════════════
   // 13b. #181 regression — encrypted backup re-enumerates profiles
