+++ tests/test-openrouter.js
@@ -286,12 +286,13 @@ assert('startup-oauth-callbacks.js clears pending OAuth state after callback',
 assert('startup-oauth-callbacks.js marks fresh OpenRouter settings local for sync',
   startupOAuthSrc.includes('markOpenRouterOAuthSettingsLocal()'));
 const syncSrc = read('js/sync.js');
+const syncApplySrc = read('js/sync-apply.js');
 assert('sync preserves fresh OpenRouter OAuth provider/key against stale pull',
-  syncSrc.includes('shouldKeepLocalOpenRouterOAuthSetting') && syncSrc.includes("'labcharts-openrouter-key'"));
+  syncApplySrc.includes('shouldKeepLocalOpenRouterOAuthSetting') && syncApplySrc.includes("'labcharts-openrouter-key'"));
 assert('sync preserves fresh local AI settings against stale pull',
-  syncSrc.includes('AI_SETTINGS_LOCAL_LOCK_UNTIL_KEY') && syncSrc.includes('shouldKeepLocalAISetting(key)'));
+  syncApplySrc.includes('AI_SETTINGS_LOCAL_LOCK_UNTIL_KEY') && syncApplySrc.includes('shouldKeepLocalAISetting(key)'));
 assert('sync refreshes AI header after remote AI settings apply',
-  syncSrc.includes('window.updateChatHeaderModel?.()') && syncSrc.includes('window.refreshWebSearchToggle?.()'));
+  syncApplySrc.includes('window.updateChatHeaderModel?.()') && syncApplySrc.includes('window.refreshWebSearchToggle?.()'));
 assert('startup sync reconciliation pushes local AI setting drift',
   syncSrc.includes('newer local AI settings') && syncSrc.includes('collectAISettings()'));
 const cssSrc = read('styles.css');
